docs: Update INSTALL
The INSTALL file contained several minor errors, typos and was generally
not up-to-date in regards to what ./configure provides today.  In
addition, several URL references have moved around to new homes.

This is just a backport of git master 6099ab67122429c0 adjusting for the
slight changes in release/2.4.  The contents is the same as git master,
as it is still valid for v2.4.

Signed-off-by: David Sommerseth <davids@openvpn.net>
Acked-by: Gert Doering <gert@greenie.muc.de>
Message-Id: <20190412202020.24045-1-davids@openvpn.net>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg18364.html
Signed-off-by: Gert Doering <gert@greenie.muc.de>